requirements.txt
accelerate==1.7.0
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.0
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
certifi==2022.12.7
chardet==5.2.0
charset-normalizer==2.1.1
click==8.2.1
cmake==3.25.0
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.2
cycler==0.12.1
dash==3.0.4
DataProperty==1.1.0
datasets==2.16.1
decorator==5.2.1
decord==0.6.0
defusedxml==0.7.1
dill==0.3.7
distro==1.9.0
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.13.1
Flask==3.0.3
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2023.10.0
ftfy==6.3.1
h11==0.16.0
HiPart==1.0.4
httpcore==1.0.9
httpx==0.27.2
huggingface-hub==0.31.4
idna==3.4
imageio==2.37.0
importlib_metadata==8.7.0
ipython==8.36.0
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.15
KDEpy==1.1.12
kiwisolver==1.4.8
lazy_loader==0.4
lit==15.0.7
loguru==0.7.3
MarkupSafe==2.1.5
matplotlib==3.10.3
matplotlib-inline==0.1.7
mbstrdecoder==1.1.4
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.15
narwhals==1.40.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
open3d==0.18.0
open_clip_torch==2.26.1
openai==1.35.3
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
packaging==25.0
pandas==2.2.3
parso==0.8.4
pathvalidate==3.2.3
patsy==1.0.1
peft==0.15.2
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.8
plotly==6.1.1
prompt_toolkit==3.0.51
propcache==0.3.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyarrow-hotfix==0.7
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
pyquaternion==0.9.9
pytablewriter==1.2.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.28.1
retrying==1.3.4
rpds-py==0.25.1
safetensors==0.5.3
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
statsmodels==0.14.4
supervision==0.21.0
sympy==1.13.3
tabledata==1.3.4
tcolorpy==0.1.7
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.15
tokenizers==0.21.1
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.3
treelib==1.7.1
triton==3.3.0
typepy==1.3.4
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.2.31
ultralytics-thop==2.0.14
urllib3==1.26.13
wcwidth==0.2.13
Werkzeug==3.0.6
widgetsnbextension==4.0.14
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
